Pull the Voice Activity Detector out from the AGC
This change generates bit-exact values when running through audioproc_f than before.

R=andrew@webrtc.org, bloch@google.com

Review URL: https://codereview.webrtc.org/1181933002.

Cr-Commit-Position: refs/heads/master@{#9465}

+++ webrtc/modules/audio_processing/agc/agc_manager_direct.cc
@@ -321,7 +321,7 @@ void AgcManagerDirect::SetCaptureMuted(bool muted) {
 }
 
 float AgcManagerDirect::voice_probability() {
-  return static_cast<float>(agc_->voice_probability());
+  return agc_->voice_probability();
 }
 
 int AgcManagerDirect::CheckVolumeAndReset() {
